+++ lib/widgets/refresh.dart
@@ -3,6 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_waya/flutter_waya.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 
+void sendRefreshType([RefreshCompletedType refresh]) =>
+    eventBus.emit(refreshEvent, refresh ?? RefreshCompletedType.refresh);
+
 class Refreshed extends StatefulWidget {
   Refreshed({
     Key key,
@@ -34,7 +37,7 @@ class Refreshed extends StatefulWidget {
 
   ///  可不传controller，
   ///  若想关闭刷新组件可以通过发送消息
-  ///  sendMessage(RefreshCompletedType.refresh);
+  ///  sendRefreshType(RefreshCompletedType.refresh);
   final RefreshController controller;
   final VoidCallback onRefresh;
   final VoidCallback onLoading;
@@ -70,35 +73,37 @@ class _RefreshedState extends State<Refreshed> {
   @override
   void initState() {
     super.initState();
+
     controller = widget.controller ?? RefreshController(initialRefresh: false);
-    Ts.addPostFrameCallback((Duration callback) => eventListen((dynamic data) {
-          if (data == null) return;
-          if (data != null && data is RefreshCompletedType) {
-            switch (data) {
-              case RefreshCompletedType.refresh:
-                controller?.refreshCompleted();
-                break;
-              case RefreshCompletedType.refreshFailed:
-                controller?.refreshFailed();
-                break;
-              case RefreshCompletedType.refreshToIdle:
-                controller?.refreshToIdle();
-                break;
-              case RefreshCompletedType.onLoading:
-                controller?.loadComplete();
-                break;
-              case RefreshCompletedType.loadFailed:
-                controller?.loadFailed();
-                break;
-              case RefreshCompletedType.loadNoData:
-                controller?.loadNoData();
-                break;
-              case RefreshCompletedType.twoLevel:
-                controller?.twoLevelComplete();
-                break;
-            }
-          }
-        }));
+    Ts.addPostFrameCallback(
+        (Duration callback) => eventBus.add(refreshEvent, (dynamic data) {
+              if (data == null) return;
+              if (data != null && data is RefreshCompletedType) {
+                switch (data) {
+                  case RefreshCompletedType.refresh:
+                    controller?.refreshCompleted();
+                    break;
+                  case RefreshCompletedType.refreshFailed:
+                    controller?.refreshFailed();
+                    break;
+                  case RefreshCompletedType.refreshToIdle:
+                    controller?.refreshToIdle();
+                    break;
+                  case RefreshCompletedType.onLoading:
+                    controller?.loadComplete();
+                    break;
+                  case RefreshCompletedType.loadFailed:
+                    controller?.loadFailed();
+                    break;
+                  case RefreshCompletedType.loadNoData:
+                    controller?.loadNoData();
+                    break;
+                  case RefreshCompletedType.twoLevel:
+                    controller?.twoLevelComplete();
+                    break;
+                }
+              }
+            }));
   }
 
   @override