+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * ExtensionString add insert void
   * add [BasisText]  [BasisTextStyle] 
   * modify [RichSpan] to [RichTextSpan]
+  * modify [sendMessage(RefreshCompletedType.refresh)] to [sendRefreshType(RefreshCompletedType.refresh)]
 ## [1.5.0]
   * change Tools to Ts.
   * change GlobalMaterial to GlobalWidgetsApp.